+++ oap-server/server-storage-plugin/storage-banyandb-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/banyandb/BanyanDBIndexInstaller.java
@@ -40,7 +40,7 @@ public class BanyanDBIndexInstaller extends ModelInstaller {
     public BanyanDBIndexInstaller(Client client, ModuleManager moduleManager, BanyanDBStorageConfig config) {
         super(client, moduleManager);
         this.config = config;
-        MetadataRegistry.INSTANCE.initializeIntervals(config.getOverrideGroupIntervals());
+        MetadataRegistry.INSTANCE.initializeIntervals(config.getSpecificGroupSettings());
     }
 
     @Override