+++ docs/en/changes/changes.md
@@ -35,9 +35,9 @@
 * Change `elasticsearch/flushInterval` to 5s(was 15s).
 * Optimize `flushInterval` of ElasticSearch BulkProcessor to avoid extra periodical flush in the continuous bulk streams. 
 * Remove measure/stream specific interval settings in BanyanDB.
-* Add global interval settings and allow to override settings for each single group in BanyanDB.
+* Add global-specific settings used to override global configurations (e.g `segmentIntervalDays`, `blockIntervalHours`) in BanyanDB.
 * Use TTL-driven interval settings for the `measure-default` group in BanyanDB.
-* Fix wrong group of non timeRelative metadata in BanyanDB.
+* Fix wrong group of non time-relative metadata in BanyanDB.
 
 #### UI